+++ src/plone/app/fhirfield/indexes/es/mapping_types.py
@@ -6,221 +6,143 @@
 # All imports here
 
 
-__author__ = 'Md Nazrul Islam <email2nazrul@gmail.com>'
+__author__ = "Md Nazrul Islam <email2nazrul@gmail.com>"
 
-date_pattern = '-?[0-9]{4}(-(0[1-9]|1[0-2])(-(0[0-9]|[1-2][0-9]|3[0-1]))?)?'  # noqa: E501
-datetime_pattern = '-?[0-9]{4}(-(0[1-9]|1[0-2])(-(0[0-9]|[1-2][0-9]|3[0-1])(T([01][0-9]|2[0-3]):[0-5][0-9]:[0-5][0-9](\\.[0-9]+)?(Z|(\\+|-)((0[0-9]|1[0-3]):[0-5][0-9]|14:00)))?)?)?'  # noqa: E501
-Bool = {
-    "type": "boolean",
-    "store": False
-}
-Float = {
-    "type": "float",
-    "store": False
-}
+date_pattern = (
+    "-?[0-9]{4}(-(0[1-9]|1[0-2])(-(0[0-9]|[1-2][0-9]|3[0-1]))?)?"
+)  # noqa: E501
+datetime_pattern = "-?[0-9]{4}(-(0[1-9]|1[0-2])(-(0[0-9]|[1-2][0-9]|3[0-1])(T([01][0-9]|2[0-3]):[0-5][0-9]:[0-5][0-9](\\.[0-9]+)?(Z|(\\+|-)((0[0-9]|1[0-3]):[0-5][0-9]|14:00)))?)?)?"  # noqa: E501
+Bool = {"type": "boolean", "store": False}
+Float = {"type": "float", "store": False}
 
-Integer = {
-    "type": "integer",
-    "store": False
-}
-Long = {
-    "type": "long",
-    "store": False
-}
+Integer = {"type": "integer", "store": False}
+Long = {"type": "long", "store": False}
 
-Token = {
-    'type': 'keyword',
-    'index': True,
-    'store': False
-}
+Token = {"type": "keyword", "index": True, "store": False}
 
-KeywordToken = {
-    'type': 'keyword',
-    'index': True,
-    'store': False
-}
+KeywordToken = {"type": "keyword", "index": True, "store": False}
 
-Text = {
-    'type': 'text',
-    'index': True,
-    'store': False,
-    'analyzer': 'standard'
-}
+PathToken = {"type": "text", "index": True, "store": False, "analyzer": "path_analyzer"}
 
-SearchableText = {
-    'type': 'text',
-    'index': True,
-    'analyzer': 'standard',
-    'store': False
-}
+Text = {"type": "text", "index": True, "store": False, "analyzer": "standard"}
+
+SearchableText = {"type": "text", "index": True, "analyzer": "standard", "store": False}
 
 Date = {
-    'type': 'date',
-    'format': 'date_time_no_millis||date_optional_time',
-    'store': False
+    "type": "date",
+    "format": "date_time_no_millis||date_optional_time",
+    "store": False,
 }
 
-Timing = {
-    'properties': {
-        'event': Date,
-        'code': Token
-    }
-}
+Timing = {"properties": {"event": Date, "code": Token}}
 
 Reference = {
-    'properties': {
-        'reference': KeywordToken
+    "properties": {
+        "reference": {
+            "type": "text",
+            "index": True,
+            "store": False,
+            "analyzer": "fhir_reference_analyzer",
+        }
     }
 }
 
 Attachment = {
-    'properties': {
-        'url': Token,
-        'language': Token,
-        'title': Text,
-        'creation': Date
-    }
+    "properties": {"url": Token, "language": Token, "title": Text, "creation": Date}
 }
 
-Coding = {
-    'properties': {
-        'system': Token,
-        'code': Token,
-        'display': Token
-    }
-}
+Coding = {"properties": {"system": Token, "code": Token, "display": Token}}
 
 CodeableConcept = {
-    'properties': {
-        'text': Text,
-        'coding': {
-            'type': 'nested',
-            'properties': Coding['properties']
-        }
+    "properties": {
+        "text": Text,
+        "coding": {"type": "nested", "properties": Coding["properties"]},
     }
 }
 
-Period = {
-    'properties': {
-        'start': Date,
-        'end': Date
-    }
-}
+Period = {"properties": {"start": Date, "end": Date}}
 Identifier = {
-    'properties': {
-        'use': Token,
-        'system': Token,
-        'value': Token,
-        'type': {
-            'properties': {
-                'text': Text
-            }
-        }
+    "properties": {
+        "use": Token,
+        "system": Token,
+        "value": Token,
+        "type": {"properties": {"text": Text}},
     }
 }
 Quantity = {
-    'properties': {
-        'value': Float,
-        'code': Token,
-        'system': Token,
-        'unit': Token
-    }
+    "properties": {"value": Float, "code": Token, "system": Token, "unit": Token}
 }
 
 Money = Quantity
-Range = {
-    'properties': {
-        'high': Quantity,
-        'low': Quantity
-    }
-}
+Range = {"properties": {"high": Quantity, "low": Quantity}}
 
 Age = Quantity
 Address = {
-    'properties': {
-        'city': Token,
-        'country': Token,
-        'postalCode': Token,
-        'state': Token,
-        'use': Token
+    "properties": {
+        "city": Token,
+        "country": Token,
+        "postalCode": Token,
+        "state": Token,
+        "use": Token,
     }
 }
 
 HumanName = {
-    'properties': {
-        'family': Token,
-        'text': Text,
-        'prefix': Token,
-        'given': Token,
-        'use': Token
+    "properties": {
+        "family": Token,
+        "text": Text,
+        "prefix": Token,
+        "given": Token,
+        "use": Token,
     }
 }
 Duration = Quantity
 
 ContactPoint = {
-    'properties': {
-        'period': Period,
-        'rank': Integer,
-        'system': Token,
-        'use': Token,
-        'value': Text
+    "properties": {
+        "period": Period,
+        "rank": Integer,
+        "system": Token,
+        "use": Token,
+        "value": Text,
     }
 }
 
-ContactDetail = {
-    'properties': {
-        'name': Token,
-        'telecom': ContactPoint
-    }
-}
+ContactDetail = {"properties": {"name": Token, "telecom": ContactPoint}}
 
-ContactDetail['properties']['telecom'].update({
-    'type': 'nested'
-})
+ContactDetail["properties"]["telecom"].update({"type": "nested"})
 
 Annotation = {
-    'properties': {
-        'authorReference': Reference,
-        'authorString': Text,
-        'text': Text,
-        'time': Date
+    "properties": {
+        "authorReference": Reference,
+        "authorString": Text,
+        "text": Text,
+        "time": Date,
     }
 }
 
 Dosage = {
-    'properties': {
-        'asNeededBoolean': Bool,
-        'asNeededCodeableConcept': CodeableConcept,
-        'doseQuantity': Quantity,
-        'doseRange': Range,
-        'site': CodeableConcept,
-        'text': Text,
-        'timing': Timing
+    "properties": {
+        "asNeededBoolean": Bool,
+        "asNeededCodeableConcept": CodeableConcept,
+        "doseQuantity": Quantity,
+        "doseRange": Range,
+        "site": CodeableConcept,
+        "text": Text,
+        "timing": Timing,
     }
 }
 
-RelatedArtifact = {
-    'properties': {
-        'type': Token,
-        'url': Token,
-        'resource': Reference
-    }
-}
+RelatedArtifact = {"properties": {"type": Token, "url": Token, "resource": Reference}}
 
 Signature = {
-    'properties': {
-        'contentType': Token,
-        'when': Date,
-        'whoReference': Reference,
-        'whoUri': Token
+    "properties": {
+        "contentType": Token,
+        "when": Date,
+        "whoReference": Reference,
+        "whoUri": Token,
     }
 }
 # Common
 Id = Token
-Meta = {
-    'properties': {
-        'versionId': Token,
-        'lastUpdated': Date,
-        'profile': Token
-    }
-}
+Meta = {"properties": {"versionId": Token, "lastUpdated": Date, "profile": Token}}